requirements.txt
pyyaml
# mysql-connector-python